+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Light/HDAdditionalLightData.cs
@@ -198,6 +198,14 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         [Range(0, 0.001f)]
         public float minFilterSize = 0.00001f;
 
+        // Improved Moment Shadows settings
+        [Range(1, 32)]
+        public int kernelSize = 5;
+        [Range(0.0f, 9.0f)]
+        public float lightAngle = 1.0f;
+        [Range(0.0001f, 0.01f)]
+        public float maxDepthBias = 0.001f;
+
         HDShadowRequest[]   shadowRequests;
         bool                m_WillRenderShadows;
         int[]               m_ShadowRequestIndices;
@@ -405,12 +413,17 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             shadowRequest.lightIndex = lightIndex;
             // We don't allow shadow resize for directional cascade shadow
             shadowRequest.allowResize = m_Light.type != LightType.Directional;
+            shadowRequest.lightType = (int) m_Light.type;
 
             // Shadow algorithm parameters
             shadowRequest.shadowSoftness = shadowSoftness / 100f;
             shadowRequest.blockerSampleCount = blockerSampleCount;
             shadowRequest.filterSampleCount = filterSampleCount;
             shadowRequest.minFilterSize = minFilterSize;
+
+            shadowRequest.kernelSize = (uint)kernelSize;
+            shadowRequest.lightAngle = (lightAngle * Mathf.PI / 180.0f);
+            shadowRequest.maxDepthBias = maxDepthBias;
         }
 
 #if UNITY_EDITOR
@@ -663,7 +676,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             // m_Light.intensity is in luminance which is the value we need for emissive color
             Color value = m_Light.color.linear * m_Light.intensity;
             if (useColorTemperature)
-                value *= LightUtils.CorrelatedColorTemperatureToRGB(m_Light.colorTemperature);
+                value *= Mathf.CorrelatedColorTemperatureToRGB(m_Light.colorTemperature);
             value.r = Mathf.Clamp01(value.r);
             value.g = Mathf.Clamp01(value.g);
             value.b = Mathf.Clamp01(value.b);